requirements/local.txt
-r base.txt
ipython==1.0.0
pytest
vcrpy>=1.0.0
